# Store objects using COPY FROM BINARY on pg8000 and psycopg2