This can be 20% or more faster.

Addresses #247 but there are still some cleanups that would be nice to do.