@jamadden 2019-06-22
| Benchmark                            | master | This branch |
|--------------------------------------|-----------|-----------------------------|
| psycopg2_hf: add 1000 objects        | 134 ms    | 99.0 ms: 1.36x faster (-26%) |
| psycopg2_hf: update 1000 objects     | 144 ms    | 104 ms: 1.39x faster (-28%)  |
| psycopg2cffi_hf: add 1000 objects    | 182 ms    | 125 ms: 1.46x faster (-32%)  |
| psycopg2cffi_hf: update 1000 objects | 183 ms    | 127 ms: 1.43x faster (-30%)  |